# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI项目管理后端API服务，提供项目管理、任务管理、文档生成等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AI Project Backend API
  version: "1.0"
paths:
...
  /api/v1/enterprises/{enterprise_id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取企业域所有可用权限列表,支持按资源过滤
      parameters:
      - description: 企业ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: 资源类型过滤 (e.g., project, task, document)
        in: query
        name: resource
        type: string
      - default: true
        description: 是否仅显示活跃权限
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出企业可用权限
      tags:
      - 企业权限管理
  /api/v1/enterprises/{enterprise_id}/roles:
    get:
      consumes:
      - application/json
      description: Get paginated list of all roles in the enterprise
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search by role name or code
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of enterprise roles
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid enterprise ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List enterprise roles
      tags:
      - enterprise-roles
    post:
      consumes:
      - application/json
      description: Create a new role within the enterprise
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: Create role request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Role already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create enterprise role
      tags:
      - enterprise-roles
...
  /api/v1/enterprises/{enterprise_id}/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of all users in the enterprise
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search by username or email
        in: query
        name: search
        type: string
      - description: Filter by status (active/inactive)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of enterprise users
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - Invalid enterprise ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List enterprise users
      tags:
      - enterprise-users
    post:
      consumes:
      - application/json
      description: Invite a new user to join the enterprise or add existing user
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: Invite user request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: User invited successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already in enterprise
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Invite user to enterprise
      tags:
      - enterprise-users
  /api/v1/enterprises/{enterprise_id}/users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove a user's access to the enterprise
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User removed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove user from enterprise
      tags:
      - enterprise-users
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific user in the enterprise
      parameters:
      - description: Enterprise ID
        in: path
        name: enterprise_id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Enterprise user details with roles
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get enterprise user
      tags:
      - enterprise-users
...
  /api/v1/system/permissions:
    get:
      consumes:
      - application/json
      description: 获取所有系统级权限列表,支持按资源和活跃状态过滤
      parameters:
      - description: 资源类型过滤 (e.g., user, role, enterprise)
        in: query
        name: resource
        type: string
      - default: true
        description: 是否仅显示活跃权限
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 权限列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出所有系统权限
      tags:
      - 系统权限管理
    post:
      consumes:
      - application/json
      description: 创建新的系统级权限定义
      parameters:
      - description: 权限信息
        in: body
        name: permission
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 创建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 无权限
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 权限代码已存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建系统权限
      tags:
      - 系统权限管理
  /api/v1/system/roles:
    get:
      consumes:
      - application/json
      description: Get list of all system roles with pagination
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search by role name or code
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: List of system roles
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - System admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List system roles
      tags:
      - system-roles
    post:
      consumes:
      - application/json
      description: Create a new system role with specified permissions
      parameters:
      - description: Create system role request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created system role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Role already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create system role
      tags:
      - system-roles
...
  /api/v1/system/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of all system users
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search by username or email
        in: query
        name: search
        type: string
      - description: Filter by status (active/inactive)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of system users
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - System admin only
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List system users
      tags:
      - system-users
    post:
      consumes:
      - application/json
      description: Create a new system user with specified roles
      parameters:
      - description: Create system user request
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created system user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create system user
      tags:
      - system-users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"